Jwt ,Security Added

+++ src/main/java/com/foodapp/springfoodapp/serviceImpl/AddressServicesImpl.java
@@ -2,9 +2,7 @@ package com.foodapp.springfoodapp.serviceImpl;
 
 
 import com.foodapp.springfoodapp.entiry.Address;
-import com.foodapp.springfoodapp.entiry.Customer;
 import com.foodapp.springfoodapp.repository.AddressRepo;
-import com.foodapp.springfoodapp.request.CreateAddress;
 import com.foodapp.springfoodapp.service.AddressServices;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,34 +16,72 @@ public class AddressServicesImpl implements AddressServices {
 
     private final AddressRepo addressRepo;
 
-
+    //todo -> error "Query did not return a unique result: 3 results were returned",
     @Override
     public List<Address> saveAddress(List<Address> addresses) {
+        return addressRepo.saveAll(addresses);
+    }
 
+    @Override
+    public Address save(Address address) {
+        Address saveAddress = addressRepo.save(address);
+        return saveAddress;
+    }
+
+    @Override
+    public List<Address> getAll() {
         return addressRepo.findAll();
     }
 
     @Override
-    public List<Address> getAddress() {
-        return addressRepo.findAll();
+    public Address findByAddressId(Integer id) {
+        Optional<Address> address = addressRepo.findById(id);
+        if (address.isEmpty()) {
+            throw new IllegalArgumentException("Address Not Found");
+        } else {
+            return address.get();
+        }
     }
 
-    @Override
-    public String deleteAddress(int addressId) {
-        addressRepo.deleteById(addressId);
-        return "delete successful";
-    }
+//    @Override
+//    public void deleteAddress(int addressId) {
+//        addressRepo.deleteById(addressId);
+//    }
 
     @Override
-    public Address updateAddress(Address updateAddress) {
-        Address address = addressRepo.findById(updateAddress.getAddressId())
+    public void deleteAddress(int addressId) {
+        Optional<Address> addressOpt = addressRepo.findById(addressId);
+        if (addressOpt.isPresent()) {
+            addressRepo.deleteById(addressId);
+        }
+    }
+
+
+    @Override
+    public Address updateAddress(Address updateAddress, int id) {
+        Address address = addressRepo.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("No Address Id"));
 
-        address.setState(updateAddress.getState());
-        address.setArea(updateAddress.getArea());
-        address.setCountry(updateAddress.getCountry());
-        address.setPinCode(updateAddress.getPinCode());
-        address.setCity(updateAddress.getCity());
+        if (address.getArea() != null) {
+            address.setArea(updateAddress.getArea());
+        }
+        if (address.getCity() != null) {
+            address.setCity(updateAddress.getCity());
+        }
+        if (address.getState() != null) {
+            address.setState(updateAddress.getState());
+        }
+        if (address.getCountry() != null) {
+            address.setCountry(updateAddress.getCountry());
+        }
+        if (address.getPinCode() != null) {
+            address.setCountry(updateAddress.getCountry());
+        }
+//        address.setState(updateAddress.getState());
+//        address.setCity(updateAddress.getCity());
+//        address.setArea(updateAddress.getArea());
+//        address.setCountry(updateAddress.getCountry());
+//        address.setPinCode(updateAddress.getPinCode());
         addressRepo.save(address);
         return address;
     }
@@ -53,11 +89,6 @@ public class AddressServicesImpl implements AddressServices {
     @Override
     public Address findByCity(String city) {
         return addressRepo.findByCity(city);
-    }
-
-    @Override
-    public Address save(Address address) {
-        return addressRepo.save(address);
     }
